+++ source/javascripts/pages/TriggersPage/TriggersPage.stories.tsx
@@ -10,6 +10,7 @@ export default {
     isWebsiteMode: false,
     pipelines: ['foo', 'bar', 'ci-test-long-name-example-with-potential-truncat-foooooooooo-very-long'],
     workflows: ['foo', 'bar'],
+    setDiscard: () => {},
   },
   argTypes: {
     onTriggerMapChange: { type: 'function' },
