+++ source/javascripts/pages/WorkflowsPage/components/WorkflowConfigPanel/SelectiveTriggering/AddTriggerDialog.tsx
@@ -11,10 +11,11 @@ type AddTriggerDialogProps = {
   onClose: () => void;
   onSubmit: (action: 'add' | 'edit', trigger: TriggerItem) => void;
   editedItem?: TriggerItem;
+  workflows: string[];
 };
 
 const AddTriggerDialog = (props: AddTriggerDialogProps) => {
-  const { isOpen, onClose, currentTriggers, onSubmit, editedItem } = props;
+  const { isOpen, onClose, currentTriggers, onSubmit, editedItem, workflows } = props;
 
   const defaultValues: FormItems = useMemo(() => {
     return {
@@ -107,7 +108,7 @@ const AddTriggerDialog = (props: AddTriggerDialogProps) => {
       >
         <DialogBody>
           <Text textStyle="body/lg/regular" color="text/secondary" marginBlockEnd="24">
-            Set up the trigger conditions that should all be met to execute the WORKFLOW NAME Workflow.{' '}
+            Set up the trigger conditions that should all be met to execute the {workflows} Workflow.{' '}
           </Text>
           {fields.map((item, index) => {
             return (
